Currency Converter
The Currency Converter project is a user-friendly web application that enables users to convert between different currencies quickly and accurately. Built using a combination of HTML, CSS, and JavaScript, this project offers a seamless and intuitive interface for performing real-time currency conversions.The project utilizes JavaScript to fetch real-time exchange rates from a reliable currency exchange API. This ensures that the conversions are accurate and up to date, reflecting the latest market rates.